+++ inbm-lib/inbm_common_lib/constants.py
@@ -24,7 +24,7 @@ CONFIG_CHANNEL = 'ma/configuration/update/'
 
 # Request constants
 CONFIG_LOAD = "load"
-CONFIG_APPEND = "append"
+CONFIG_REMOVE = "remove"
 
 # Source
 LOCAL_SOURCE = 'local'